Mount hostfs
Opening /dev/tpm0 required routing through to the enclave host. Mounting
the filesystem onto hostfs routes all file options through the host.
This sample currently routes everything to the host, but we could
probably limit this to just /dev/tpm0 to make it more secure so other
file operations are not 'leaked' to the host.
BIG NOTE! The TPM stores secrets! Clearly we need a secure session to be
set up before we do any real work. This sample will be updated in time
to achieve this. Currently I will be happy to get any simple operation
to work!