Moving to newer version of openenclave openssl lib
Previously this project was using an early version of the OpenSsl
library we built for use within an SGX enclave. This in turn was
dependent on work in progress file IO work within open enclave itself.
Now we have more official work for all of these, this project now pulls
in the official features.