Updated tpm2_tss build to work with latest tpm2-tss
Apparently a few changes happened that I missed in the previous check-in
when moving to the latest tpm2-tss. The specification of what device to
use for the TPM changed, as well as log level build flags.
I will make a check in that helps to detect this better in a future
check in.
The planets are back in alignment and the build works again.